feat: supervisor agent overview banner + fix 60 UI TypeScript errors
Feature: Supervisor Agent Overview (管理者 Agent 专属展示)
- agents.ts: add renderSupervisorBanner() — full banner for supervisor
  (team name, status dot, member avatars x4+overflow, pause/resume);
  compact read-only banner for sub-members (team name + view button)
- agents.ts: extend renderAgentOverview() with 5 team context params
- app.ts: fix agentsPanel type — add "chat" | "outputs" to union
- app-render.ts: fix blank page on supervisor click — capture prevPanel
  before guard so control-flow narrowing does not suppress chat/outputs
  preload branches
- agents.css: add .supervisor-banner and .supervisor-banner--member CSS
- en.ts / zh-CN.ts: add team.action.viewTeam, supervisorBanner keys

Fix: UI TypeScript errors 60 -> 0
- app-gateway.ts: add 4 update-system fields to GatewayHost interface;
  fix .then() callback param type (unknown + inner cast)
- views/overview.ts: extend OverviewProps with 35 missing fields
  (usage, models, security props + 13 callbacks)
- types.ts: add SkillsMarketResponse.message?; add
  TeamProjectDetail.visibility.displayEmoji?
- app-chat.ts + app-view-state.ts: add voiceMode? to handleSendChat opts
- app.ts: add lastUpgradeResult:null to licenseState initializer;
  fix renderApp() cast via unknown
- app-render.ts: add onDeviceRemove to second nodesProps block;
  cast t and orchestratorState at orchestrator entry points
- app-settings.ts: fix closeOrchestrator cast via unknown
- views/playground.ts: replace t() with tMaybe() for dynamic keys
- i18n en.ts + zh-CN.ts: add 21 missing keys across channels, sessions,
  config.field.channels.feishu/dingtalk/wecom sections
- test fixes: error-hints, chat-modality-guard, model-config,
  chat.security, team-projects (type corrections + missing fields)

+++ ui/src/ui/views/chat.security.test.ts
@@ -3,7 +3,7 @@
  * Chat 视图安全测试 - 附件 ID 生成
  */
 
-import { describe, expect, it, vi } from "vitest";
+import { beforeEach, describe, expect, it, vi } from "vitest";
 
 // Mock crypto.getRandomValues
 const originalGetRandomValues = crypto.getRandomValues;
@@ -88,7 +88,7 @@ describe("generateAttachmentId Security Tests", () => {
     // Verify Uint8Array was passed
     const call = getRandomValuesSpy.mock.calls[0];
     expect(call[0]).toBeInstanceOf(Uint8Array);
-    expect(call[0].length).toBe(6);
+    expect((call[0] as Uint8Array).length).toBe(6);
 
     getRandomValuesSpy.mockRestore();
   });
